test(cgcignore): make the CLI cases hermetic and move them to the integration suite (#1422)
The TC-12+ cases shared one fixed /tmp/cgc_test/_home database across every
test and run, wiped it mid-suite, and were gated on a live Neo4j — so they
were permanently skipped on most machines and intermittently failed with a
different name each full-suite run wherever they did run. Each test now gets
a fresh HOME and an explicit embedded backend (no server, no gate, no wipe),
and the subprocess-driven cases live in tests/integration where that cost
belongs. tc01-tc11 (pure PathSpec logic) stay in the unit suite.
